+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -4,6 +4,7 @@ import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
@@ -62,6 +63,10 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable Long id) {
-        userService.deleteUser(id);
+        if (id > 0) {
+            userService.deleteUser(id);
+        } else {
+            throw new NotFoundException("Пользователь с id " + id + " не существует");
+        }
     }
 }
